requirements.txt
streamlit
opencv-python
numpy
ultralytics
deep-sort-realtime
torch
torchvision
Pillow
mtcnn
facenet-pytorch
fastapi
uvicorn
python-multipart
requests
python-dotenv